created role repository, syntax changes to SubProjectRepository, moved employee methods from SubProjectRepository to EmployeeRepository
