+++ src/main/java/com/example/estimatea/repository/mapper/SubProjectEmployeeMapper.java
@@ -1,6 +1,5 @@
 package com.example.estimatea.repository.mapper;
 
-import com.example.estimatea.model.Employee;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 
@@ -8,12 +7,12 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 @Component
-public class SubProjectEmployeeMapper implements RowMapper<Employee> {
+public class SubProjectEmployeeMapper implements RowMapper<Integer> {
 
     @Override
-    public Employee mapRow(ResultSet rs, int rowNum) throws SQLException {
-        Employee employee = new Employee();
-        employee.setEmployeeId(rs.getInt("employee_id"));
-        return employee;
+    public Integer mapRow(ResultSet rs, int rowNum) throws SQLException {
+        int employeeId = rs.getInt("employee_id");
+
+        return employeeId;
     }
 }
